Created a shared utils path and moved FormatDate util to dateUtils. Also, fixed imports
