+++ src/components/StatusBar/components/StatusSection.tsx
@@ -9,8 +9,9 @@ import {
 } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { useFormContext } from "../../Contexts/FormContext";
-import { FormatDate, SortHistory } from "../utils";
+import { SortHistory } from "../utils";
 import { StatusIconMap } from "../icons";
+import { FormatDate } from "../../../utils";
 
 const StyledAvatar = styled(Avatar)({
   background: "transparent",
